requirements.txt
game-sdk
python-dotenv
requests
tweepy
google-genai
chromadb>=1.0.0
twikit
